+++ spring-mvc/src/main/java/cz/fi/muni/pa165/springmvc/config/MySpringMvcConfig.java
@@ -42,8 +42,6 @@ public class MySpringMvcConfig extends WebMvcConfigurerAdapter {
     public void addViewControllers(ViewControllerRegistry registry) {
         log.debug("mapping URL / to home view");
         registry.addViewController("/").setViewName("home");
-        log.debug("mapping URL /login to login view");
-        registry.addViewController("/login").setViewName("login");
     }
 
     /**
